deepagent-test/pyproject.toml
[project]
name = "deepagent-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pydantic-ai-backend>=0.0.3",
    "pydantic-ai-todo>=0.1.1",
    "pydantic-deep>=0.2.9",
]
